♻️ refactor(quality-phase2): extract chat history parsing from ChatPanel
The 110-line history→Message[] transformation (Converse-format block
parsing, toolResult re-linking, attachment markers) moves to
chatHistory.ts as pure functions. ChatPanel drops 649→547 lines and the
transformation gains 8 unit tests covering both Local and Cloud formats.
