+++ web-ui/src/components/chat/ChatPanel.tsx
@@ -26,6 +26,7 @@ import { Send, ChevronRight } from "lucide-react"
 import { ModeSelector } from "./ModeSelector"
 import { usePreferences } from "@/hooks/usePreferences"
 import { notifyError } from "@/lib/errors"
+import { isLocalHistoryFormat, parseLocalHistory, parseCloudHistory } from "./chatHistory"
 
 interface ChatPanelProps {
   deckId: string
@@ -176,114 +177,11 @@ export const ChatPanel = forwardRef<ChatPanelHandle, ChatPanelProps>(function Ch
         const history = await getChatHistory(sessionId, idToken ?? "", deckId || undefined)
         if (history.length > 0) {
           // Local mode: .chat.json is already in ChatPanel's internal format
-          if (IS_LOCAL && (history[0] as unknown as Record<string, unknown>)?.toolUses !== undefined) {
-            stream.setMessages(history.map((m) => {
-              const raw = m as unknown as Record<string, unknown>
-              return {
-                role: ((raw.role as string) || "assistant") as "user" | "assistant",
-                content: (typeof raw.content === "string" ? raw.content : "") as string,
-                toolUses: (raw.toolUses as ToolUse[]) || [],
-                blocks: (raw.blocks as ({ type: "text"; text: string } | { type: "tool"; tool: ToolUse })[]) || undefined,
-              }
-            }))
+          if (IS_LOCAL && isLocalHistoryFormat(history)) {
+            stream.setMessages(parseLocalHistory(history))
             return
           }
-          const parsed: Message[] = []
-          for (const m of history) {
-            let text = ""
-            const toolUses: ToolUse[] = []
-            const snippets: { label: string; text: string }[] = []
-
-            if (typeof m.content === "string") {
-              text = m.content.replace(/<!--sdpm:[^>]*-->\n?/g, "")
-            } else if (Array.isArray(m.content)) {
-              const contentBlocks = m.content as unknown as Record<string, unknown>[]
-              if (m.role === "user" && contentBlocks.some((b) => b.toolResult)) {
-                for (const block of contentBlocks) {
-                  const b = block
-                  if (b.toolResult) {
-                    const tr = b.toolResult as Record<string, unknown>
-                    const tuId = tr.toolUseId as string
-                    const status = (tr.status as string) || "success"
-                    let resultText = ""
-                    for (const c of (tr.content as Record<string, unknown>[]) || []) {
-                      if (c.text) resultText += c.text as string
-                    }
-                    if (parsed.length > 0) {
-                      const prev = parsed[parsed.length - 1]
-                      if (prev.role === "assistant") {
-                        const matchedTool = prev.toolUses.find((t) => t.toolUseId === tuId)
-                        if (matchedTool) {
-                          matchedTool.status = status as "success" | "error"
-                          try { matchedTool.result = JSON.parse(resultText) } catch { matchedTool.result = resultText as unknown as Record<string, unknown> }
-                        }
-                        if (prev.blocks) {
-                          for (const bl of prev.blocks) {
-                            if (bl.type === "tool" && bl.tool.toolUseId === tuId) {
-                              bl.tool.status = status as "success" | "error"
-                              try { bl.tool.result = JSON.parse(resultText) } catch { bl.tool.result = resultText as unknown as Record<string, unknown> }
-                            }
-                          }
-                        }
-                      }
-                    }
-                  }
-                }
-                continue
-              }
-
-              for (const block of contentBlocks) {
-                const b = block
-                if (b.toolUse) {
-                  const tu = b.toolUse as Record<string, unknown>
-                  toolUses.push({
-                    toolUseId: (tu.toolUseId as string) || "",
-                    name: (tu.name as string) || "",
-                    input: (tu.input as Record<string, unknown>) || {},
-                  })
-                } else if (b.text && toolUses.length === 0) {
-                  const cleaned = (b.text as string).replace(/<!--sdpm:[^>]*-->\n?/g, "")
-                  if (cleaned) text += (text ? "\n" : "") + cleaned
-                }
-              }
-            }
-            if (!text.trim() && toolUses.length === 0) continue
-            const blocks: ({ type: "text"; text: string } | { type: "tool"; tool: ToolUse })[] = []
-            if (m.role === "assistant" && Array.isArray(m.content)) {
-              for (const block of m.content) {
-                const b = block as Record<string, unknown>
-                if (b.text) {
-                  blocks.push({ type: "text", text: b.text as string })
-                } else if (b.toolUse) {
-                  const tu = b.toolUse as Record<string, unknown>
-                  blocks.push({ type: "tool", tool: {
-                    toolUseId: (tu.toolUseId as string) || "",
-                    name: (tu.name as string) || "",
-                    input: (tu.input as Record<string, unknown>) || {},
-                  }})
-                }
-              }
-            }
-            parsed.push({
-              role: m.role as "user" | "assistant",
-              content: text,
-              toolUses,
-              blocks: blocks.length > 0 ? blocks : undefined,
-              snippets: snippets.length > 0 ? snippets : undefined,
-              ...((m.role === "user") && (() => {
-                const attRe = /\[Attached:\s*(.+?)\s*\(uploadId:\s*[^)]+\)\]/g
-                const atts: { fileName: string; fileType: string }[] = []
-                let am: RegExpExecArray | null
-                while ((am = attRe.exec(text)) !== null) {
-                  const fn = am[1]
-                  const ext = fn.split(".").pop()?.toLowerCase() || ""
-                  const mimeMap: Record<string, string> = { pptx: "application/vnd.openxmlformats-officedocument.presentationml.presentation", pdf: "application/pdf", png: "image/png", json: "application/json", md: "text/markdown", txt: "text/plain" }
-                  atts.push({ fileName: fn, fileType: mimeMap[ext] || "application/octet-stream" })
-                }
-                return atts.length > 0 ? { attachments: atts } : {}
-              })()),
-            })
-          }
+          const parsed = parseCloudHistory(history)
           if (parsed.length > 0) stream.setMessages(parsed)
         }
       } finally {
